requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bpy==3.6.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Cython==3.1.6
datasets==4.4.1
dill==0.4.0
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.3.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torch-geometric==2.7.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
